+++ src/core/Table/Table.test.tsx
@@ -6,7 +6,7 @@ import { fireEvent, render, screen } from '@testing-library/react';
 import React from 'react';
 import { Table, TableProps } from './Table';
 import * as IntersectionHooks from '../utils/hooks/useIntersection';
-import { TableFilters } from './Filters';
+import { TableFilters } from './filters';
 
 const intersectionCallbacks = new Map<Element, () => void>();
 jest
