+++ src/core/Table/filters/TextFilter/TextFilter.tsx
